created updateView method to update game view for each command. updated enum userAction to be in MotionsController. updated sensorHandler.handle to use a double instead of an int

+++ inwatchz/app/src/main/java/com/example/sarahwada/myapplication/sensorhandler/ActionEventListener.java
@@ -19,7 +19,7 @@ public abstract class ActionEventListener implements SensorEventListener {
         this.sensorManager = sensorManager;
     }
 
-    public boolean startListener(int duration) {
+    public boolean startListener(double duration) {
         sensorManager.registerListener(this, this.sensor, SensorManager.SENSOR_DELAY_NORMAL);
         return true;
     }
